forwprop: Don't do copy-prop-aggregates from statements that could throw [PR120599]
In the testcase provided, currently we lose the landing pad for the exception that could
throw from the aggregate load as we remove one copy and the second statement where load
happens was not marked as throwable before so the landing pad for that internal throw is
now gone.

The fix is to ignore statements that could throw (internally or externally).

	PR tree-optimization/120599

gcc/ChangeLog:

	* tree-ssa-forwprop.cc (optimize_agr_copyprop): Don't try to copy
	from statements that throw.

gcc/testsuite/ChangeLog:

	* g++.dg/torture/noncall-eh-1.C: New test.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>