gcn: Add 'nops' insn, extend comments
Use 's_nops' with a number instead of multiple of 's_nop' when
manually adding 1 to 5 wait state. This helps with
the instruction cache and helps a tiny bit with PR119367 where
a two-byte variable overflows in the debugging location view handling.

Add a comment about 'sc0' to TARGET_GLC_NAME as for atomics it is
unrelated to the scope but to whether the result is stored; i.e.
using e.g. 'sc1' instead of 'sc0' will have undesired consequences!

Update the comment above print_operand_address to document 'R' and 'V';
those are used below as "Temporary hack.", but it makes sense to see
them in the list.

gcc/ChangeLog:

	* config/gcn/gcn-opts.h (enum hsaco_attr_type): Add comment
	about 'sc0'.
	* config/gcn/gcn.cc (gcn_md_reorg): Use gen_nops instead of gen_nop.
	(print_operand_address): Document 'R' and 'V' in the
	pre-function comment as well.
	* config/gcn/gcn.md (nops): Add.