output: Move an special # (256) to a new macro
This is a followup to the review of mergability of CSWTCH patch
located at https://gcc.gnu.org/pipermail/gcc-patches/2025-July/690810.html.
Moves the special # (256) to a macro so it is not used bare in the source
and there is only the need to change it in one place.
This special # was added with r0-37392-g201556f0e00580 which added the original mergeable
section support to gcc.

Pushed as obvious after build and test on x86_64.

gcc/ChangeLog:

	* output.h (MAX_ALIGN_MERGABLE): New define.
	* tree-switch-conversion.cc (switch_conversion::build_one_array):
	Use MAX_ALIGN_MERGABLE instead of 256.
	* varasm.cc (mergeable_string_section): Likewise
	(mergeable_constant_section): Likewise

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>