tree-optimization/120687 - avoid disturbing reduction chains in reassoc
Reassoc carefully ranks operands to form reduction chains for
vectorization so we are careful to not apply any width related
changes in the early pass.  Unfortunately we are not careful
enough.  The following gates fma related re-ordering and also
the >= 3 ops tail "optimization" which is the culprit here.

This does not fix the reported inefficient vectorization when
using signed integer reductions yet.

	PR tree-optimization/120687
	* tree-ssa-reassoc.cc (reassociate_bb): Do not disturb
	the sorted operand order in the early pass.
	* tree-vect-slp.cc (vect_analyze_slp): Dump when a detected
	reduction chain fails SLP discovery.

	* gcc.dg/vect/pr120687-1.c: New testcase.
	* gcc.dg/vect/pr120687-2.c: Likewise.