testsuite: Make aarch64/cmpbr.c more forgiving
The 8-bit and 16-bit tests in cmpbr.c assumed an inverted operand
order ("w1, w0"), but it's possible to use the uninverted operand
order too.  This patch generalises the tests to support both forms.

This is a prerequisite for a later patch that adds a new
simplify-rtx.cc rule.

gcc/testsuite/
	* gcc.target/aarch64/cmpbr.c: Support both operand orders
	for 8-bit and 16-bit comparisons.