IFCVT: Fix factor_out_operators correctly for more than 1 phi [PR121295]
r16-2590-ga51bf9e10182cf was not the correct fix for this in the end.
Instead a much simplier and localized fix is needed, just change the phi
that is being worked on with the new result and arguments that is from the
factored out operator.
This solves the issue of not having result in the IR and causing issues that way.

Bootstrapped and tested on x86_64-linux-gnu.
Note this depends on reverting r16-2590-ga51bf9e10182cf.

	PR tree-optimization/121236
	PR tree-optimization/121295

gcc/ChangeLog:

	* tree-if-conv.cc (factor_out_operators): Change the phi node
	to the new result and args.

gcc/testsuite/ChangeLog:

	* gcc.dg/torture/pr121236-1.c: New test.
	* gcc.dg/torture/pr121295-1.c: New test.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>