[x86] factor out worker from ix86_builtin_vectorization_cost
The following factors out a worker that gets a mode argument
rather than a vectype argument.  That makes a difference when
we hit the fallback in add_stmt_cost for scalar stmts where
vectype might be NULL and thus mode is derived from the scalar
stmt there.  But ix86_builtin_vectorization_cost does not
have access to the stmt.  So the patch instead dispatches
to the new ix86_default_vector_cost there, passing down the mode
we derived from the stmt.

This is to avoid regressions with a patch that makes even more
scalar stmt costings have a vectype passed.

	* config/i386/i386.cc (ix86_default_vector_cost): Split
	out from ...
	(ix86_builtin_vectorization_cost): ... this and use
	mode instead of vectype as argument.
	(ix86_vector_costs::add_stmt_cost): Call
	ix86_default_vector_cost instead of ix86_builtin_vectorization_cost.