asf: Fix null pointer dereference in is_store_forwarding [PR121303]
We were calling `is_store_forwarding` with a NULL value for `off_val`,
which was causing a null pointer dereference in `is_constant`, leading
to an ICE.

This patch updates the call to `is_constant` in `is_store_forwarding`
and adds a check for `off_val`, in order to update it with the right
value.

Bootstrapped/regtested on AArch64 and x86_64.

	PR rtl-optimization/121303

gcc/ChangeLog:

	* avoid-store-forwarding.cc (is_store_forwarding): Add check
	for `off_val` in `is_store_forwarding`.

gcc/testsuite/ChangeLog:

	* gcc.target/i386/pr121303.c: New test.