c++: improve non-constant template arg diagnostic
A conversation today pointed out that the current diagnostic for this case
doesn't mention the constant evaluation failure, it just says e.g.

"'p' is not a valid template argument for 'int*' because it is not the
address of a variable"

This patch improves the diagnostic in C++17 and above (post-N4268) to
diagnose failed constant-evaluation.

gcc/cp/ChangeLog:

	* pt.cc (convert_nontype_argument_function): Check
	cxx_constant_value on failure.
	(invalid_tparm_referent_p): Likewise.

gcc/testsuite/ChangeLog:

	* g++.dg/tc1/dr49.C: Adjust diagnostic.
	* g++.dg/template/func2.C: Likewise.
	* g++.dg/cpp1z/nontype8.C: New test.